check that max_inactivity_time is defined before using it as timeout

+++ src/libs/data-staging/DataDeliveryComm.cpp
@@ -226,7 +226,7 @@ namespace DataStaging {
     }
     // check for stuck child process (no report through comm channel)
     Arc::Period t = Arc::Time() - last_comm;
-    if (t >= transfer_params.max_inactivity_time) {
+    if (transfer_params.max_inactivity_time > 0 && t >= transfer_params.max_inactivity_time) {
       logger_->msg(Arc::ERROR, "Transfer killed after %i seconds without communication", t.GetPeriod());
       child_->Kill(1);
       delete child_;
